[UID2] Use UID2 service in non-authenticated flow
The unauthenticated flow also uses the UID2 service now. It is done by
generating a random Guid and requesting the UID2 token for it. In this way,
we make sure there is a different UID2 token for every new anonymous user.
UID2 does not verify if the input is a valid email, we may need to change
this behaviour in the future. However, it's enough for the PoC for now.

Change-Id: I529f936ef68bd0b47344bd1c3125441fd699d83b
JIRA:USRF-50